--- com.devsquad10.product/src/main/java/com/devsquad10/product/infrastructure/config/RedisCachingConfig.java
@@ -1,34 +1,0 @@
-package com.devsquad10.product.infrastructure.config;
-
-import java.time.Duration;
-
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.cache.CacheKeyPrefix;
-import org.springframework.data.redis.cache.RedisCacheConfiguration;
-import org.springframework.data.redis.cache.RedisCacheManager;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.serializer.RedisSerializationContext;
-import org.springframework.data.redis.serializer.RedisSerializer;
-
-@Configuration
-@EnableCaching
-public class RedisCachingConfig {
-
-	@Bean
-	public RedisCacheManager cacheManager(RedisConnectionFactory redisConnectionFactory) {
-
-		RedisCacheConfiguration configuration = RedisCacheConfiguration
-			.defaultCacheConfig()
-			.disableCachingNullValues()
-			.entryTtl(Duration.ofSeconds(120)) // 기본 유지 시간 120초
-			.computePrefixWith(CacheKeyPrefix.simple())
-			.serializeValuesWith(RedisSerializationContext.SerializationPair.fromSerializer(RedisSerializer.java()));
-
-		return RedisCacheManager
-			.builder(redisConnectionFactory)
-			.cacheDefaults(configuration)
-			.build();
-	}
-}
